+++ tests/testthat/helper.R
@@ -1,0 +1,3 @@
+testthat::set_state_inspector(function() {
+  getAllConnections()
+})